Retry ChannelMessage if processed between channel close and ShutdownSignal
There exists a small window in which a channel could be closed, but the
ChannelActor state indicates otherwise. This makes the "dropIfNoChannel"
flag for ChannelMessage unreliable, as it WILL drop in this case.

Since it's possible that the ChannelMessage is responsible for breaking
and closing the channel, we don't want to retry it forever, so a limited
retry mechanism is introduced. 4 failures in a row with a successful
reconnect between each is a good indication of a problematic
ChannelMessage, after which point the ChannelMessage will be dropped.

If dropIfNoChannel == true, then retry mechanism is not invoked.